backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=2.0
psycopg2-binary
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-auth
requests
pytesseract
pdf2image
Pillow
redis
rq
pydantic
python-docx
pandas
openpyxl
xlrd
apscheduler
cryptography
PyPDF2
